move models to parser, change generic
