+++ src/parsers/__init__.py
@@ -1,4 +1,2 @@
 # flake8: noqa
-from parsers.command_parser import CommandParser, ChatCommandParser
-from parsers.arg_schema import ArgSchema
-from parsers.command_info import CommandInfo
+from parsers.command_parser import CommandParser, ChatCommandParser, ArgSchema, CommandInfo
